removing unnesserary as casts

+++ rust-doomgeneric/engine/src/am_map.rs
@@ -1275,7 +1275,7 @@ pub unsafe fn AM_drawWalls(state: &mut GameState) {
                     AM_drawMline(state, &raw mut l, TSWALLCOLORS + lightlev);
                 }
             }
-        } else if (*state.g_game.player_mut(state.am_map.plr)).powers[PowerType::pw_allmap as i32 as usize] != 0 {
+        } else if (*state.g_game.player_mut(state.am_map.plr)).powers[PowerType::pw_allmap as usize] != 0 {
             if (*li).flags as i32 & LINE_NEVERSEE == 0 {
                 AM_drawMline(state, &raw mut l, GRAYS + 3 as i32);
             }
@@ -1381,7 +1381,7 @@ pub unsafe fn AM_drawPlayers(state: &mut GameState) {
             && PlayerId(i as u8) != state.am_map.plr)
         {
             if state.g_game.playeringame[i as usize] {
-                if (*p).powers[PowerType::pw_invisibility as i32 as usize] != 0 {
+                if (*p).powers[PowerType::pw_invisibility as usize] != 0 {
                     color = 246 as i32;
                 } else {
                     color = their_colors[their_color as usize];
